Refactor q8 schedule telemetry module
